fix: 折りたたみドロワーをpersistentに変更 #199
スクロール不具合解消のため
